refactor: 좋아요 박스 순수한 컴포넌트로 리팩토링
