--- src/components/molecule/ToggleBox/ToggleBox.stories.jsx
@@ -1,9 +1,0 @@
-import { ToggleBox } from '@components';
-
-export default {
-  title: 'Component/Domain/ToggleBox',
-};
-
-export const Default = () => {
-  return <ToggleBox />;
-};
